Move seed module into coco
These seeds are specific to the coco domain and therfore are you used in
that context exclusively.
