+++ proxy/src/lib.rs
@@ -40,5 +40,4 @@ pub mod keystore;
 mod notification;
 pub mod project;
 pub mod registry;
-pub mod seed;
 pub mod session;
