+++ proxy/src/main.rs
@@ -4,12 +4,12 @@ use librad::paths;
 use radicle_keystore::pinentry::SecUtf8;
 
 use proxy::coco;
+use proxy::coco::seed;
 use proxy::config;
 use proxy::env;
 use proxy::http;
 use proxy::keystore;
 use proxy::registry;
-use proxy::seed;
 use proxy::session;
 
 /// Flags accepted by the proxy binary.
@@ -46,13 +46,13 @@ async fn main() -> Result<(), Box<dyn std::error::Error>> {
             });
 
             client
-        },
+        }
         host => {
             let host = url17::Host::parse(host)?;
             radicle_registry_client::Client::create_with_executor(host)
                 .await
                 .expect("unable to construct devnet client")
-        },
+        }
     };
 
     let temp_dir = tempfile::tempdir().expect("test dir creation failed");
